test(ScreeningSeatView): 나눠져 있는 Test 파일들을 관련있는 하나의 파일로 모으기
